requirements.txt
aiohttp==3.8.4
fastapi==0.92.0
pydantic==1.10.5
python-dotenv==0.21.1
revChatGPT==2.3.4
uvicorn==0.20.0
